create a helper class for validating InvertedIndex.searchIndex parameters
